feat(tests): enhance fixture descriptions and add new integration tests
- Updated the description for the 'corrupt' fixture in `build-fixture-manifest.ts` and `manifest.json` to clarify that incomplete games are dropped during parsing and not counted in skipped games.
- Revised the README to reflect changes in test focus areas, including updates to `fixtures.test.ts` and `filters.test.ts` for better clarity on test coverage.
- Introduced new integration tests in `filters.test.ts`, `multi-run.test.ts`, and `trackers.test.ts` to validate filtering, multi-run functionality, and tracker behaviors, ensuring comprehensive coverage of new features.
