+++ test/helpers/fixtures.ts
@@ -27,7 +27,7 @@ export interface FixtureEntry {
     file: string;
     description: string;
     expected: { games: number; moves: number; skippedGames?: number };
-    analyzeOptions?: { onError?: 'abort' | 'skip-game' };
+    analyzeOptions?: { onError?: 'abort' | 'skip-game'; workers?: false };
     golden?: {
         tileTracker?: TileTrackerGolden;
     };
